requirements.txt
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
Django==2.2.10
djangorestframework==3.12.4
idna==2.10
iniconfig==1.1.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.5
